Competed encrypt. Partial Decrypt and Keygen
Have a working encrypt of all modes and saving to individual files.
There is a commented section in the ECB encryption section that will
decrypt the file and works properly.

Decrypt is populated but has hard coded key and IV that must be
replaced if you rerun the encrypt program. Decrypt doesn't properly
decrypt the ECB file and crashes on the CBC file complaining about
invalid padding. Problem seems to be with the key which is stored as a
string but needs to be a binary array for the decrypt streamtransform.
this problem could go away if the read/write binary key capability was
completed.

Keygen now writes a hexencoded keyfile as well as a binary file.
Changed the key type from byte array to secblockbyte following the
example at (http://www.cryptopp.com/wiki/Advanced_Encryption_Standard)