Decrypt Cleaned up but without IV preservation
This Decrypt now matches the encrypt in form and function. The IV is
not being passed yet, I think we can use a stringsink to build a
concatonation of the IV and the cipher and reverse the process in the
decrypt. Not sure how the string variable will hadle large files though.